SMAbstract completed
