+++ src/algmatch/stableMatchings/stableMarriageProblem/smManOptimal.py
