+++ src/algmatch/stableMatchings/stableMarriageProblem/smPreferenceInstance.py
@@ -1,0 +1,22 @@
+"""
+Store preference lists for Stable Marriage stable matching algorithm.
+"""
+
+from abstractClasses.abstractPreferenceInstance import AbstractPreferenceInstance
+from stableMatchings.stableMarriageProblem.fileReader import FileReader
+from stableMatchings.stableMarriageProblem.dictionaryReader import DictionaryReader
+
+
+class SMPreferenceInstance(AbstractPreferenceInstance):
+    def __init__(self, filename: str | None = None, dictionary: dict | None = None) -> None:
+        super().__init__(filename, dictionary)
+
+    def _load_from_file(self, filename: str) -> None:
+        reader = FileReader(filename)
+        self.men = reader.men
+        self.women = reader.women
+
+    def _load_from_dictionary(self, dictionary: dict) -> None:
+        reader = DictionaryReader(dictionary)
+        self.men = reader.men
+        self.women = reader.women
