+++ src/algmatch/stableMatchings/stableMarriageProblem/smWomanOptimal.py
